move backup printer helpers to backup_printer.go
